wip - some cleanup
