--- _archive/001_Build_v01/config/assemblyai-gwf-intel-82de817e0155.json
@@ -1,12 +1,0 @@
-{
-  "type": "service_account",
-  "project_id": "assemblyai-gwf-intel",
-  "private_key_id": "82de817e01555905ee5253d759ca360d1b092d11",
-  "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQDIvBPpDfah59Sr\nGKmgmVBFK3AfJImkpDgC44tM7Mr8J1JIDGEuDHDyQZKlyenvVwPVWX0sdxqQfEPf\nWnc9pMbRfekaF6V4iep/GjgSgXMIaqHeP1UjblmdxRkyqw2VJFHYGTODk19+DqCQ\npNFve5QIqggigPhqtA34YPcIPq62EKngWPpndYK7mHNO3uS1n95MjV3fExFHXQOs\nya5H5nPz7w/a6e8GY8FQKrLawz7zMToeFPHb7IlpVWGWX43j/F3Y28JZycopJVlX\nHQBruQkPa8DkBhdBvpWdnidGNTDLVUDYf6HjOwhwe9vQarW8fAI0Z73oYhUlFiTF\n2swPAVIrAgMBAAECggEAYSOW1QCTWe1I3pVcPOIEVLy9EO0LCNgJXOiFWC/82MIk\nkMYg3vpZS/qMf1szMEXhDoU0TGjEu6fO5o6uRm/vYKB4Rc6sniTLjD+BY/8tziRf\n+Z+XUUKM7orbc0gQlhxR2Ov1IFaYbvWgIO8JhNP+IqU5s4zK6kQAwLIXyW8f9+1D\n1zi8Ht11PqlfQLxPZTV2seGrnaqBzJR5aoGohjSoZI1bJQZOzFYmMcQdSM/u0hBk\nHCIN4gLENeNbLtRHb/FtacP8n1Uhz8rDzBdVcNHLvnnw5sghNhG6sf1QTeAe5uEO\nsfd19urfVnyxtCW0r0wj0d9LlkrSulxa6T4kykT9pQKBgQDscmHGRtHuGelbhEIN\nooIjQhSQxvTS2Gdm/KKVP7aK5lK84//+9wY3KKV8sj4l27l0u8BSFmO5T+Gwv/6E\n8+lbdhkcLLJvDai1Y927ikFDTRGrmSESK3Co0aDYhi0hqRLAxSICuRJA/ig1tLaH\ngTZtV+wIe09pMGVgdklEGmO8pwKBgQDZVaoKWp+qLtJ+U0y2UoLH7NMO9YtxQhjT\nhoddSkFRHrWOPQDncBzsRDKfTxs/orVc9+RA4F9n2yv2VgObc4ECA042D7HmzWS+\nFJmDkynRBHfhy5pI/9tJwJ62m3IXRTpjJ0YqPhQMT2MDLQLOtAXEZ8Tdn77piSeH\nOqQuY52a3QKBgQDrsGy0Bz4VeHC7j8+D5dyRJrp3QksirSINb4DlAviD5mjM7vzr\n+vKQIw0qikYW2Lo5mTlw2e6X9v+Cfy9G+hc/2y3WfttyhYwJG74Xxg2d94HIqkBj\n6zdGoIcTrlPofK87xoVLb425iV8tgbB8zBkjz4NxRPF4agxzphUMNL38ewKBgEAm\nGNqVhSsmRtsimzH1iMSaEnIH5/C2Ib3Np4bRtBXNiqlBOFMXtHHIb2i2YMSXshCO\n2sCLeZ2eA4viN45auWlusFckqs4FMgco7Wb4uQ8I0pnDDvcdDgzm3fHoT/Sual15\n6ivtxt0bUY5OatW4LTubpXAFBsTHxvMTTAKlRI/BAoGBALj2wAW4xwQ82ag/mHQ/\nNCwmOOvKNP26Wbip7cDZct0istCTZEIit+tEhoYXXEPDJvpJz6JubChpu7TQ7d3M\nuhHJqfgcGuB7dWmBu0EdRHdHFfmqwcXSddqLtpS9EjMC7Bmgw1gEQ4Dtfp3hMn6P\nT5WwSzsganVrYQjFr8yUuOw8\n-----END PRIVATE KEY-----\n",
-  "client_email": "assemblyai-gwf@assemblyai-gwf-intel.iam.gserviceaccount.com",
-  "client_id": "107674456083453179324",
-  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
-  "token_uri": "https://oauth2.googleapis.com/token",
-  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
-  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/assemblyai-gwf%40assemblyai-gwf-intel.iam.gserviceaccount.com"
-}
