--- _archive/001_Build_v01/config/client_secret_607980291111-sdul4bik3snfjn2hci8k9rjqjo8lmf0p.apps.googleusercontent.com.json
@@ -1,1 +1,0 @@
-{"installed":{"client_id":"607980291111-sdul4bik3snfjn2hci8k9rjqjo8lmf0p.apps.googleusercontent.com","project_id":"assemblyai-gwf-intel","auth_uri":"https://accounts.google.com/o/oauth2/auth","token_uri":"https://oauth2.googleapis.com/token","auth_provider_x509_cert_url":"https://www.googleapis.com/oauth2/v1/certs","client_secret":"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","redirect_uris":["http://localhost"]}}
