--- _archive/001_Build_v01/versions/v_01-simpleTranscript-working/main.py
@@ -1,8 +1,0 @@
-import sys
-from api_communication import *
-
-
-filename = sys.argv[1]
-audio_url = upload(filename)
-
-save_transcript(audio_url, filename)
