--- _archive/001_Build_v01/versions/v_02/api_secrets.py
@@ -1,1 +1,0 @@
-API_KEY_ASSEMBLYAI = "57ea7ea3435d40208709cf48c720f121"
