--- _archive/001_Build_v02/requirements.txt
@@ -1,10 +1,0 @@
-certifi==2022.9.24
-charset-normalizer==2.1.1
-idna==3.4
-pip==21.3.1
-requests==2.28.1
-setuptools==60.2.0
-tqdm==4.64.1
-urllib3==1.26.12
-wheel==0.37.1
-XlsxWriter==3.0.3
